chore: standardize imports/exports of /style dir
