+++ libs/spark/src/index.ts
@@ -139,7 +139,7 @@ export { default as makeStyles } from './makeStyles';
 
 export { default as styled } from './styled';
 
-export { theme } from './styles';
+export * from './styles';
 
 export { default as useTheme } from './useTheme';
