+++ libs/spark/src/styles/index.ts
@@ -1,1 +1,1 @@
-export { theme } from './theme';
+export { default as theme } from './theme';
