+++ libs/spark/src/styles/theme.ts
@@ -1,9 +1,9 @@
-import { createTheme } from '@material-ui/core';
-import { overrides } from './overrides';
-import { palette } from './palette';
-import { props } from './props';
-import { shadows } from './shadows';
-import { typography } from './typography';
+import createTheme from '@material-ui/core/styles/createTheme';
+import overrides from './overrides';
+import palette from './palette';
+import props from './props';
+import shadows from './shadows';
+import typography from './typography';
 
 const initialTheme = createTheme({
   palette,
@@ -12,7 +12,9 @@ const initialTheme = createTheme({
   typography,
 });
 
-export const theme = createTheme({
+const theme = createTheme({
   ...initialTheme,
   overrides: overrides(initialTheme),
 });
+
+export default theme;
