+++ libs/spark/src/styles/typography.ts
@@ -1,4 +1,3 @@
-import * as CSS from 'csstype';
 import {
   TypographyOptions,
   TypographyUtils,
@@ -129,7 +128,7 @@ const customVariants: Record<SparkVariant, TypographyStyle> = {
   'code-sm': buildVariant(500, 14, 20, undefined, undefined, codeFontFamily),
 };
 
-export const typography: SparkTypographyOptions = {
+const typography: SparkTypographyOptions = {
   // override default Roboto
   fontFamily: defaultFontFamily,
   // override default 14px
@@ -153,113 +152,4 @@ export const typography: SparkTypographyOptions = {
   ...customVariants,
 };
 
-const nunitoRegular: CSS.AtRule.FontFace = {
-  fontFamily: '"Nunito"',
-  fontStyle: 'normal',
-  fontDisplay: 'swap',
-  fontWeight: 400,
-  src: [
-    'url(/fonts/nunito-regular-webfont.woff2) format("woff2")',
-    'url(/fonts/nunito-regular-webfont.woff) format("woff")',
-  ].join(', '),
-};
-
-const nunitoItalic: CSS.AtRule.FontFace = {
-  fontFamily: '"Nunito"',
-  fontStyle: 'italic',
-  fontDisplay: 'swap',
-  fontWeight: 400,
-  src: [
-    'url(/fonts/nunito-italic-webfont.woff2) format("woff2")',
-    'url(/fonts/nunito-italic-webfont.woff) format("woff")',
-  ].join(', '),
-};
-
-const nunitoSemibold: CSS.AtRule.FontFace = {
-  fontFamily: '"Nunito"',
-  fontStyle: 'normal',
-  fontDisplay: 'swap',
-  fontWeight: 600,
-  src: [
-    'url(/fonts/nunito-semibold-webfont.woff2) format("woff2")',
-    'url(/fonts/nunito-semibold-webfont.woff) format("woff")',
-  ].join(', '),
-};
-
-const nunitoSemiboldItalic: CSS.AtRule.FontFace = {
-  fontFamily: '"Nunito"',
-  fontStyle: 'italic',
-  fontDisplay: 'swap',
-  fontWeight: 600,
-  src: [
-    'url(/fonts/nunito-semibolditalic-webfont.woff2) format("woff2")',
-    'url(/fonts/nunito-semibolditalic-webfont.woff) format("woff")',
-  ].join(', '),
-};
-
-const nunitoBoldItalic: CSS.AtRule.FontFace = {
-  fontFamily: '"Nunito"',
-  fontStyle: 'italic',
-  fontDisplay: 'swap',
-  fontWeight: 700,
-  src: [
-    'url(/fonts/nunito-bolditalic-webfont.woff2) format("woff2")',
-    'url(/fonts/nunito-bolditalic-webfont.woff) format("woff")',
-  ].join(', '),
-};
-
-const nunitoBold: CSS.AtRule.FontFace = {
-  fontFamily: '"Nunito"',
-  fontStyle: 'normal',
-  fontDisplay: 'swap',
-  fontWeight: 700,
-  src: [
-    'url(/fonts/nunito-bold-webfont.woff2) format("woff2")',
-    'url(/fonts/nunito-bold-webfont.woff) format("woff")',
-  ].join(', '),
-};
-
-const nunitoExtraboldItalic: CSS.AtRule.FontFace = {
-  fontFamily: '"Nunito"',
-  fontStyle: 'italic',
-  fontDisplay: 'swap',
-  fontWeight: 800,
-  src: [
-    'url(/fonts/nunito-extrabolditalic-webfont.woff2) format("woff2")',
-    'url(/fonts/nunito-extrabolditalic-webfont.woff) format("woff")',
-  ].join(', '),
-};
-
-const nunitoExtrabold: CSS.AtRule.FontFace = {
-  fontFamily: '"Nunito"',
-  fontStyle: 'normal',
-  fontDisplay: 'swap',
-  fontWeight: 800,
-  src: [
-    'url(/fonts/nunito-extrabold-webfont.woff2) format("woff2")',
-    'url(/fonts/nunito-extrabold-webfont.woff) format("woff")',
-  ].join(', '),
-};
-
-const sourceCodeProRegular: CSS.AtRule.FontFace = {
-  fontFamily: '"Source Code Pro"',
-  fontStyle: 'normal',
-  fontDisplay: 'swap',
-  fontWeight: 500,
-  src: [
-    'url(/fonts/sourcecodepro-medium-webfont.woff2) format("woff2")',
-    'url(/fonts/sourcecodepro-medium-webfont.woff) format("woff")',
-  ].join(', '),
-};
-
-export const fontFaces = [
-  nunitoRegular,
-  nunitoItalic,
-  nunitoBold,
-  nunitoBoldItalic,
-  nunitoExtrabold,
-  nunitoExtraboldItalic,
-  nunitoSemibold,
-  nunitoSemiboldItalic,
-  sourceCodeProRegular,
-];
+export default typography;
